+++ swap/src/network/transport.rs
@@ -6,11 +6,13 @@ use libp2p::core::{identity, Transport};
 use libp2p::dns::TokioDnsConfig;
 use libp2p::mplex::MplexConfig;
 use libp2p::noise::{self, NoiseConfig, X25519Spec};
+use libp2p::websocket::WsConfig;
 use libp2p::{yamux, PeerId};
 use std::time::Duration;
 
 /// Builds a libp2p transport with the following features:
 /// - TcpConnection
+/// - WebSocketConnection
 /// - DNS name resolution
 /// - authentication via noise
 /// - multiplexing via yamux or mplex
@@ -22,8 +24,10 @@ pub fn build(id_keys: &identity::Keypair) -> Result<SwapTransport> {
 
     let tcp = TokioTcpConfig::new().nodelay(true);
     let dns = TokioDnsConfig::system(tcp)?;
+    let websocket = WsConfig::new(dns.clone());
 
-    let transport = dns
+    let transport = websocket
+        .or_transport(dns)
         .upgrade(Version::V1)
         .authenticate(noise)
         .multiplex(SelectUpgrade::new(